requirements.txt
Flask==2.2.5
numpy==1.24.3
pandas==1.5.3
tensorflow==2.12.0
keras==2.12.0
opencv-python==4.7.0.72
Pillow==9.5.0
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
imagehash==4.3.1
tqdm==4.65.0
shutilwhich==1.1.0
